First try at auto deploying postgres database Added DATABASE_URL detection from env variables Added CreateUser method to db package -> Creates user in database Fixed datatypes for models

+++ perms/get.go
@@ -36,7 +36,7 @@ func GetVisibleProjects(User models.User) ([]models.Project, error) {
 }
 
 //GetVisibleQueuesFromProject returns all queues visible for the given user
-func GetVisibleQueuesFromProject(User models.User, ProjectID int) ([]models.Queue, error) {
+func GetVisibleQueuesFromProject(User models.User, ProjectID int64) ([]models.Queue, error) {
 	Perms, err := CombinePermissions(User)
 	if err != nil {
 		return make([]models.Queue, 0), err
